LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/interaction/focus/focus-keyboard-js-expected.txt



FAIL focus-keyboard-js assert_equals: #text should be focused by our event listener expected Element node <input type="text" id="text" value="abc"></input> but got Element node <body><input type="text" id="text" value="abc">
<script>
...
